moving implementation back to main module. removing redundant regex flags.
